phase-3/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
asyncpg==0.30.0
PyJWT==2.9.0
python-dateutil==2.9.0
pydantic-settings==2.6.0
alembic==1.14.0
psycopg2-binary==2.9.10
cryptography>=41.0.0
openai-agents==0.2.9

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.10.0
ruff==0.7.4